fix(amazon-s3): enhance upload file & authentication
