+++ packages/pieces/amazon-s3/package.json
@@ -1,4 +1,4 @@
 {
   "name": "@activepieces/piece-amazon-s3",
-  "version": "0.2.0"
+  "version": "0.3.0"
 }
